LPS-56013 Rename packages and classes for consistency
