+++ modules/portal-security/portal-security-auto-login/src/com/liferay/portal/security/auto/login/basic/auth/header/settings/BasicAuthHeaderAutoLoginCompanyServiceSettingsIdMapping.java
@@ -1,0 +1,40 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.auto.login.basic.auth.header.settings;
+
+import com.liferay.portal.kernel.settings.definition.SettingsIdMapping;
+import com.liferay.portal.security.auto.login.basic.auth.header.configuration.BasicAuthHeaderAutoLoginConfiguration;
+import com.liferay.portal.security.auto.login.basic.auth.header.constants.BasicAuthHeaderAutoLoginConstants;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Tomas Polesovsky
+ */
+@Component
+public class BasicAuthHeaderAutoLoginCompanyServiceSettingsIdMapping
+	implements SettingsIdMapping {
+
+	@Override
+	public Class<?> getConfigurationBeanClass() {
+		return BasicAuthHeaderAutoLoginConfiguration.class;
+	}
+
+	@Override
+	public String getSettingsId() {
+		return BasicAuthHeaderAutoLoginConstants.SERVICE_NAME;
+	}
+
+}
