+++ portal-impl/src/portal.properties
@@ -3379,7 +3379,7 @@
     # Set this to true to automatically import users from LDAP if they do not
     # exist in the portal. The property "auto.login.hooks" must contain a
     # reference to the class
-    # com.liferay.portal.security.auto.login.requestheader.RequestHeaderAutoLogin
+    # com.liferay.portal.security.auto.login.request.header.RequestHeaderAutoLogin
     # to enable request header authentication.
     #
     #request.header.auth.import.from.ldap=false
@@ -3611,7 +3611,7 @@
     # This property is not read by the portal except for portal properties
     # overridden by liferay-hook.xml. It remains here only as a reference.
     #
-    #auth.verifier.pipeline=com.liferay.portal.security.auth.verifier.basic.BasicAuthHeaderAuthVerifier,com.liferay.portal.security.auth.verifier.DigestAuthenticationAuthVerifier,com.liferay.portal.security.auth.verifier.parameter.ParameterAuthVerifier,com.liferay.portal.security.auth.verifier.PortalSessionAuthVerifier,com.liferay.portal.security.auth.verifier.TunnelingServletAuthVerifier
+    #auth.verifier.pipeline=com.liferay.portal.security.auth.verifier.basic.auth.header.BasicAuthHeaderAuthVerifier,com.liferay.portal.security.auth.verifier.DigestAuthenticationAuthVerifier,com.liferay.portal.security.auth.verifier.request.parameter.RequestParameterAuthVerifier,com.liferay.portal.security.auth.verifier.PortalSessionAuthVerifier,com.liferay.portal.security.auth.verifier.TunnelingServletAuthVerifier
 
     #
     # Each authentication verifier can have custom properties set via the
@@ -3646,10 +3646,10 @@
     #auth.verifier.DigestAuthenticationAuthVerifier.urls.includes=N/A
 
     #
-    # ParameterAuthVerifier
+    # RequestParameterAuthVerifier
     #
-    #auth.verifier.ParameterAuthVerifier.hosts.allowed=
-    #auth.verifier.ParameterAuthVerifier.urls.includes=N/A
+    #auth.verifier.RequestParameterAuthVerifier.hosts.allowed=
+    #auth.verifier.RequestParameterAuthVerifier.urls.includes=N/A
 
     #
     # PortalSessionAuthVerifier
